feat: Move library into own folder and added reusable hook
