+++ lib/betterRules/BetterRuleRecord.ts
